+++ nilearn/_utils/__init__.py
@@ -16,7 +16,7 @@ from nilearn._utils.helpers import (
 from .cache_mixin import CacheMixin
 from .docs import fill_doc
 from .logger import compose_err_msg
-from .niimg import _repr_niimgs, copy_img, load_niimg
+from .niimg import _repr_niimgs, load_niimg
 from .niimg_conversions import (
     check_niimg,
     check_niimg_3d,
@@ -215,7 +215,6 @@ __all__ = [
     "compare_version",
     "compose_err_msg",
     "concat_niimgs",
-    "copy_img",
     "fill_doc",
     "load_niimg",
     "remove_parameters",
