+++ nilearn/image/image.py
@@ -1391,3 +1391,26 @@ def largest_connected_component_img(imgs):
         )
 
     return ret[0] if single_img else ret
+
+
+def copy_img(img):
+    """Copy an image to a nibabel.Nifti1Image.
+
+    Parameters
+    ----------
+    img: image
+        nibabel SpatialImage object to copy.
+
+    Returns
+    -------
+    img_copy: image
+        copy of input (data, affine and header)
+    """
+    if not isinstance(img, nibabel.spatialimages.SpatialImage):
+        raise ValueError("Input value is not an image")
+    return new_img_like(
+        img,
+        safe_get_data(img, copy_data=True),
+        img.affine.copy(),
+        copy_header=True,
+    )
